+++ VersionHistory.md
@@ -11,6 +11,7 @@ Prefix the description of the change with `[major]`, `[minor]` or `[patch]` in a
 ### 0.3.0
 
 * Support `--nuget-output` with .NET builds.
+* Use the `xmldocmd` .NET Core Global Tool to generate documentation.
 
 ### 0.2.0